shop content left column refactor
