+++ components/shop-content/left-column/categories-color-indicator/component.tsx
