+++ components/shop-content/left-column/categories/list-element/component.tsx
@@ -1,0 +1,13 @@
+export default function CategoriesListElement({
+  label,
+  quantity,
+}: {
+  label: string;
+  quantity: string;
+}) {
+  return (
+    <li className="list_element">
+      <div>{label}</div> <p>{quantity}</p>
+    </li>
+  );
+}
